change opt -> short bursts and reflow short bursts to be more like tilted
